modern3: repair remaining Python numeric boundaries (R64-4/R64-P)
* insert-state scratch dtype '<u8,<u1,<u4' -> '<u8,<u1,<u8': f2 sequence IDs
  are transient construction state (deleted within the run); removes the
  2^32-read cap.  No persisted contract touched.
* MSBWTGen.py: explicit '<u8' for FM-start arrays and cumulative fmDeltas
  accumulation (no platform-inferred int32 under NumPy<2).
* MultiStringBWT pure-Python rank paths: uint64+int64 no longer promote to
  float64 (inexact above 2^53).
* TranscriptBuilder: trackPairs reader opened nonexistent 'abt.npy'; every
  writer generation creates 'about.npy' (dormant legacy bug).
* preprocessFastqs/Bams/Fastas reject >255 input files instead of silently
  wrapping about.npy uint8 file IDs.

+++ packages/msbwt-modern3/MUS/MultiStringBWT.py
@@ -350,7 +350,11 @@ class MultiStringBWT(BasicBWT):
         if (binID << self.bitPower) == index:
             ret = self.partialFM[binID][sym]
         else:
-            ret = self.partialFM[binID][sym] + np.bincount(self.bwt[binID << self.bitPower:index], minlength=6)[sym]
+            # M3-R64-W5: uint64 + int64 would promote to float64
+            # (exact only below 2^53); keep integer arithmetic exact.
+            ret = int(self.partialFM[binID][sym]) + int(
+                np.bincount(self.bwt[binID << self.bitPower:index],
+                            minlength=6)[sym])
         return int(ret)
         
     def getFullFMAtIndex(self, index):
@@ -370,7 +374,10 @@ class MultiStringBWT(BasicBWT):
         if binID << self.bitPower == index:
             ret = self.partialFM[binID]
         else:
-            ret = self.partialFM[binID] + np.bincount(self.bwt[binID << self.bitPower:index], minlength=6)
+            # M3-R64-W5: avoid float64 promotion (see getOccurrenceOfCharAtIndex).
+            ret = (self.partialFM[binID].astype('<i8') +
+                   np.bincount(self.bwt[binID << self.bitPower:index],
+                               minlength=6))
         return ret
     
     def createKmerProfile(self, k, profileCsvFN):
